Ported bug fixes
Put in a workaround for a Safari 8.0.8 bug which resulted in a select
element not updating its exposed selected option when selectedIndex is
updated programatically under some very specific conditions. Removed
the date format placeholder from ui:dateField XSLT as it (a) causes
issues in NVDA and (b) does not really reflect the range of allowed
input options.