CSS Build Sass mandatory
Removed the build of raw CSS files from build-css.xml. All CSS must be
Sass in SCSS format in the Sass directory of the wcmponents-theme
source tree or a descendant implementation.

The build now generates pattern based platform and browser CSS based on
filenames (this is a simple extension of the ieX.css builder). The IE
and platform CSS build lists are now passed into the CSS loader
JavascriptModule which removes one of the last manual build-time
convergence problems.