Update scss-lint rules
This update includes renaming all Sass variables and most of the mixins.

Turned on the NameFormat rule after making the variable and mixin name
changes.

Reduced the allowed depth of NestingDepth and SelectorDepth but allowed
some local overrides which may be revisited in future.

Made the internal documentation of the scss-lint config a bit more
readable.

Updated the SASS directory README.md to reflect these changes. Moved
some documentation around to make the documentation of XSLT and
JavaScript source directories more consistent - the XSLT README was
converted to markdown and one was added to the JavaScript wc directory.