Removed unused dependency
The wcomponent-theme-parent POM held a dependency on apache batik which
has not been used for some time. This dependency was removed. Whilst
checking for residual mentions found that the wcomponent-theme
build-images.xml retained references to an excluded svg file which has
also not existed for many years so these excludes were also removed.