Update maven site appearance
Updated the site logo to include the text “WComponents” and made the
logo a little larger. Removed the bannerRight as it is not required in
this site format and was merely a placeholder transparent PNG. Updated
the skin to something a little more modern and rather more
configurable. Added a fork on GitHub ribbon.