XSLT documentation prep
Remove old xsl documentation file xslDoc.xsl. This was a marked XML
comment parser that expected the content of particular comments to be
in APT format. Has not been used for a long time and was never a great
option.

Updated build_strip_debug.xsl to strip all elements in the doc:
namespace in preparation for adding proper xslt documentation. NOTE:
this could be done in build_all.xsl so that the documentation is not in
the uncompressed XSLT either.