Theme: CSS, SVGs, mobile
Removed all of the *.mob.scss files and rolled the CSS into the core
file. Removed the .mob.css build and JavaScript loader.

Moved almost all of the remaining SVG images back to PNG.

Made the image names more consistent - fixes #112

The upshot of all this is the Sass should be a bit easier to override,
is a bit smaller and should be a bit more friendly to iOS9 (which is a
bit sucky at the moment).