Remove debug XSLT
First step to removing the xsl:include which kills Safari on iOS
described in #126. We have always intended to move this to JavaScript
so the death of Safari merely makes this more important. The debug mode
XSLT was heavy and not completely failsafe anyway so not sad to see it
go.