OSX a11y improvement
OSX 11 Safari and Chrome both use a fixed pixel size button font. This
appears to have been done **only** to make buttons take on that cure
rounded outline appearance do beloved by Apple (but not so beloved that
they can be bothered to implement it in a way which scales). The fixed
font size can have sub-optimal impact on a11y and so we inherit button
font-family, font-size and line-height on Macs.

As a consequence we are able to remove the safari specific SCSS which
did this font inheritance for table pagination buttons.