Allow widths as strings
Update width of Column in WColumn, ColumnLayout and WTableColumn and
width of WDialog to be a string instead of an int. Deprecate
setWidth(int) in WColumn, WDialog and WTableColumn.

Update WDialog to allow height as a string (CSS dimension) instead of
an int. Deprecate setHeight(int).

Aside: fixed a pile of JavaDoc errors in wcomponents-examples which
caused build to fail if JavaDoc was required as part of building this
project (such as when first compiling after a version increment).