Bootstrapping multiFileUpload
Fixed a bootstrapping issue in multiFileUpload caused by Firefox not
focusing the file upload before a click event on the browse button. The
fix is to make sure the file upload is initialised during the click
event. This has the overhead of an extra function call but the
initialisation function is self-testing so the overhead is negligible.

Fixes #245