Table a11y, XSLT maintenance
Fixed a potential (Chrome reported but probably false negative) a11y
issue in transform of tables.

The issue was that when a table had @selectAll=‘text’ or
@expandAll=‘true’ then the thead would contain a tr consisting of one
or two td elements with appropriate colspans before a tr containing the
appropriately scoped th elements.

The a11y testing tool reported that the table did not have appropriate
column headings. It did, they just were not in the first row of the
thead and this should be OK.

To remove the report/‘fix’ the ‘issue’ I have moved the text mode
select all/none controls and expand all/none controls. They are now
part of the caption element. I am not entirely happy with this either
though I can justify it if anyone is interested:

1. these controls do form part of the table and control the table’s
components directly;
2. it could be argued that these controls form part of the description
of the table and if a table has selectable or expandable rows then it
**should** include a relevant caption explaining the purpose of these
(as they are not standard facets of a HTML table);

We deal with the span of control of the select all/none and expand
all/none adequately using aria-controls and so we _could_ put them
anywhere. I am happy to listen to arguments as to why they should be in
the table’s wrapper div and not in the table itself (I am half of this
mind myself). The argument for them being part of a caption is a bit
tenuous, especially if the table does not have any other caption
content which may provide context for these controls.

I also added a caption option to WTableOptionsExample.

The second part of this commit is a change to XSLT which creates an
out-of-viewport span element. I have modified these to all use a helper
template so any modifications to the offscreen span can happen in one
place. The helper template used to only be used by artefacts in the
table.