Prepare for removal of WFilterControl
`WFilterControl` is an orphan component which should _not_ have made it
out into the OS version of WComponents. It is a controller for an
aspect of `WDataTable` and does not have any effect on `WTable`.

`WFilterControl` has been deprecated for many release. WDataTable has
been deprecated for **years** and support is maintained for a _single_
consumer. This last user of `WDataTable` does not use `WFilterControl`.

Not only does `WFilterControl` _only_ work on `WDataTable`; it does not
work well at that. It was built for a specific use case and is not
extensible in its current form. Ajax filtering is far more flexible and
fit for multiple use cases.

WFilterControl will be deleted from the next major release. These
changes remove default client support for WFilterControl as this
support has a measurable performance impact on all tables due to some
rather heavy XSLT variable calculations.

The code to implement `WFilterControl` has been left in but commented
out. The JavaScript module `wc/ui/filterControl` has a header comment
which outlines the changes a custom theme would require to re-instate
support for `WFilterControl`.