Clean up issues from previous sub-row toggle
The first implementation of sub-row toggle had the row itself as the
state controller for its sub-rows and so needed a shed.show observer to
reset its state when its sub-rows were made visible (client mode row
expansion). This observer is no longer needed.

In addition changes were made to fix #356. This includes renaming
`activateClick` to `activateTrigger` since we are not activating a
click event.