Menu accessibility
Had to make rather a lot of changes in the menu modules to fix the a11y
issues outlined in #358.

* The aria-expanded attribute has been moved to the correct element for
menus (left where it was for trees).
* A method to acquire the expandable element was added to
wc/ui/menu/core (and overridden for trees).
* The Sass for menus other than trees has been updated.
* A dummy menu item (hidden from all) has been added to empty AJAX
powered sub-menus.
* The tree-specific pre-insertion document fragment manipulation method
was beefed up to correctly set the treeitem and group roles and
attributes and to remove the dummy element required by menus (this is
invoked when we do not have a WMenu Type context i.e. inserting a
WSubMenu as part of an ajax load of the content of another WSubMenu).
* Re-wrote the menu state writer for expanded states since role menu is
not a scoped role of role menu or menubar so we cannot use the standard
methods for getting a group within a menu.