Combo with autocomplete="list"
## WSuggestions

Amended WSuggestions to allow the autocomplete method to be set to
“both” (the default) or “list”.

When set to “both” the combo works exactly as it has previously and any
content entered which does not match a suggestion is retained. This is
the default state and still allows an application to use custom
validation to report errors to the user.

When set to “list” the suggestions will be set based on user input but
will not allow acceptance of a value not in the suggestion list.

* If the user selects from the suggestions or types a value equal to a
suggestion then that entry is retained.

If the user types content which results in at least one suggestion but
that suggestion is not the equivalent of the user content when the user
leaves the field then the first such match is deemed to be the selected
option and the textbox is updated accordingly.

* If the user enters content which results in no valid matches then
leaves the textbox then the textbox is emptied.

* If the user types content which results in at least one suggestion
then defocuses the application, refocuses the application at that field
then rapidly types new (non-matching) content and leaves the field
before the suggestion list updates then the text field will be emptied
as if there was no match.

Fixes #379.

## A11y
Fixed an a11y error which occurred in comboboxes with ajax-based
suggestions when there were no suggestions.