Include makeCommonClass in commonAttributes
* All calls to commonAttributes were followed (usually immediately) by
a call to makeCommonClass so I merged them with an allowance to call
with no additional class and to not call by using the default value of
the xsl:param. I set this default value to 0 (number) as this is not a
className which should ever be used by itself!
* Fixed a typo in the schemadoc for ui:definitionlist.