Style changes
* Changed the scss-config.yaml to remove one of our unusual rules. I
have removed the requirement for the file extension on import. This was
originally done to make import autocomplete work better in NetBeans but
a better option is to make NetBeans behave in the more commonly
accepted manner.
* Split up wc.common.buttons.scss to move rules for selectors
`.wc_btn_nada`, `.wc_btn_link` and `.wc_btn_icon` into their own files
for easier override.
* Added the class `wc_btn_nada` to elements which previously had
`wc_btn_link` so that we have _only_ one rule to remove button styles
and a separate rule to add link-like styles.
* Fixed a flaw in WSelectToggle’s XSLT which resulted in an empty
pseudo-label if the WSelectToggle is associated with a WLabel.