Remove AJAX ambiguity
WTree can have internal AJAX via get  (on item expansion) and it is an
AjaxTrigger so can POST (on item selection). This lead to an error when
bith were enabled caused by ajaxRegion.requestLoad doing an ancestor
search whilst hunting for a trigger. To fix this added an
ignoreAncestor arg to requestLoad which is passed into getTrigger.