Fix "missing" media players #504
The cause of the apparent disappearance of the audio after AJAX was
because controls was being set to NONE so it was no longer visible.

When Controls is PLAY_PAUSE, however, the same effect was happening
because the addition of the play button was not being undertaken after
insertion by AJAX.

To fix this:
1. an ajax subscriber was added to undertake initialisation work which
is needed on a component-by-component basis;
2. the play/pause button creation has been moved back to XSLT;
3. only media component with `@controls=‘play’` are pushed through the
setUp for manipulating play/pause buttons improving initialisation
performance.

Fixes #504