IDs replace '-' with '_'
There was a risk of ID clashes between theme generated IDs for
sub-components and IDs set using `setIdName(String)`. This has been
removed by changing the theme’s ID separator convention from `-` to `_`.

Found another buf in Vaadin-sass-compiler so had to change the respond
mixins to work around it and change all the calls.

Fixed a bug in XSLT which prevented client side validation and caused
an a11y error in WTextField, WEmailField and WPhoneNumberField.