Responsive improvements
DialogFrame and Calendar were improved to make the JS and CSS more
compatible. Fixed several issues with dialogs on small screens. Fixed a
heinous issue with both dialogs and calendars on tablets where they
would go **bad** if the device was switched between portrait and
landscape with a dialog or calendar open (and very amusing when both
were open).

Made some minor improvements to the draggable, positionable and
resizeable JS interfaces.