Simplify GridLayout Sass and XSLT
It would be close to impossible to retain a backwards-compatible
implementation of GridLayout and make it abstractly responsive. These
changes remove a pile on inline CSS for most common GridLayout
implementations to make it possible for applications to attempt
responsive grids.

A corollary of this (in order to maintain consistency) is that hgap and
vgap have been changed **everywhere** so that inconsistent,
unresponsive and potentially inaccessible pixel based spacing has been
replaced by class-invoked CSS rem spacing in all cases and pixel to
class conversion is determined in the theme. The XSLT provides for ’n’
classes (default 4 for each) and the Sass provides the relevant gaps.
All of this is rather light weight so should be reasonably easy to
override both in a theme and in an application.

Most of the changes in this commit are a flow-on effect of this
rationalisation as it required re-implementing all gaps.

Fixes #571.

Changed the iconography of combo box fields to a background image
instead of a pseudo-element on the wrapper. Fixes #574