re-implement web analytics #398
Added an empty overridable template and an example of a simple
implementation of web analytics using Google Analytics.

This sample implementation is in XSLT and will attach the Google
Analytics code if the application has tracking enabled and has a client
ID. Of course, to work with Google Analytics the clientId added in
WComponents has to be a Google Analytics property ID.

Fixes #398.

A corollary to this: we never implemented complex component level
tracking attributes. This is better done at the application level using
custom HTML class attributes, `WTemplate` and `addJsFile()`. Therefore
I have removed the unused skeleton schema for the complex tracking
element.

Removal of the complex tracking element allows simplification of the
schema for ui:analytic.