Responsive tables, bug fixes
* Added Sass to provide initial implementation of responsive tables. I
have chosen to key this off of a custom class applied to WTable rather
than to all tables by default. The class is `wc-respond`. Fixes #671.
* Fixed errors in list of ignored roles in ariaAnalog which could
result in false negatives in determining the nearest active ARIA
widget. Fixes #670.
* Updated CSS which could cause some labels to fall under their control
on narrow views such that a WLabel for a WCheckBox or WRadioButton will
now. Fixes #672
* Removed unused unconventional class attribute value from
WCheckBoxSelect/WRadioButtonSelect option lists as part of #651.