Rem sizes, classNames, Sass variables
* Moved all rem sizes to variables fixes #685.
* Ensured all class names follow wc- or wc_ naming convention fixes #651
* As an extension to the previous two points all non-calculation Sass
vars have been moved to _common.scss and the color vars files and any
extraneous vars files have been deleted. Any component specific vars
files which contain over-rideable mixins have been retained but
re-named to the mixin convention. This has been combined with ensuring
that all over-rideable vars are marked default. This _should_ ease the
burden of overriding vars in themes. Also rationalised the sectioning
content header sizes as part of the rem fix and moved these to a var.
These have shown to be problematic in practice and had resulted in
inconsistent header sizes and excessively complex over-rides in current
consuming themes.
* As part of #651 and for future enhancement I changed the following
class names for common helper classes to move them from the “private”
convention to the “public” convention:
  * `wc_off` for out of viewport to `wc-off`
  * `wc_invite` to turn on over/focus effects to `wc-invite`
  * `wc_btn_nada` for a button with no button-like appearance to
`wc-nobutton`
  * `wc_btn_link` for a button with a link-like appearance to
`wc-linkbutton`
* Fixed a bug in GridLayout caused by a class-name clash. This
rationalisation of class names and variables _should_ help reduce the
incidence of this in future but we still have no consistent pattern for
classes in the “private” convention so more work will need to be done.