Add wc-icon class
Added utility classes `wc-icon` and `wc-icon-right`.

`wc-icon` is used internally to reduce the amount of generated Sass. It
can be used as an add-on class value to add an FA icon to other
components. Updated the WButtonExample to show this in action.

`wc-icon-after` puts the icon to the right of the content by floating
the `before` pseudo-element. This is because FA uses the `before` and
does not allow us to easily move it to `after`.

Updated the Examples furniture to use `wc-icon` rather than using pngs
in the image buttons to select an example, refresh and reset.