Margin to class name
Changed the handling of Margin to use a limited set of class values as
per hgap/vgap/gap.

This was done to improve responsive design arrangements _and_ to better
enforce standard design requirements. Margin was pixel based for
historical reasons and does not really work in modern applications.
Also addresses aspects of #682, #689