Add new util class
Added utility class `wc-hscroll` to add horizontal scrolling facility
to components which overflow. This has been a much requested feature of
WTable but we will not turn it on by default as horizontal scrolling of
tables is usually a bad idea. **Use with caution**.

Added enum value for this class to HtmlClassUtil.