Table Sass
Rationalised the table Sass some more. My aim here is to get it as
simple as possible given the constraints of possible nested tables with
different properties. I have also merged all structural Sass into a
single file leaving the colours and icon content in separate files.

I have updated the colour definitions to allow more table colours to be
defined directly rather than using generic colours.

I have also added a mechanism to turn off CSS blocks by setting a
colour to -1. This currently applies only to table Sass but I will
extend it through the rest of the Sass over time. The point of this is
to allow a theme to turn off a chunk of colour rather than having to
override it (which could result in some nasty overrides in some cases)
or replace an entire Sass file.

Merged all table ie8 Sass into a single file.