theme build: add explicit dependency
We are getting intermittent build failures due to out of date
dependencies in upstream dependencies. This is most explicitly packages
requiring grunt ~0.4.0.

I have added grunt ~0.4.0 as an explicit dependency. This is less than
ideal but npm is pretty rubbish at dependency management and it is all
we can do at the moment.

The main culprit seems to be getusermedia-js so I will take a fork of
it and see if we can up the dependencies to >= rather than ~.