Loading indicator
Made the full screen loading indicator work in IE11. It had become
displaced. It is a bit more complex than it needs to be to solve some
IE11 issues.

Updated some aspects of aria-busy Sass whilst I was there to get a more
consistent result. The styles now have a better effect on replaced
content and text node children of busy elements.

This finishes #822 and addresses more of #689.