Implement "safe" colour classes
We have had a few requests to expose theme colours as part of the CSS
API. I am somewhat loathe to do this as it may undermine efforts to
manage consistent branding and accessibility. As a
meet-somewhere-my-side-of-the-middle compromise I have created classes
which may be used to set colours in a theme by extension and which, by
default, colour the :before pseudo element to allow for simple coloured
icons.

These classes reflect the message states:
* `wc-csuccess`
* `wc-cinfo`
* `wc-cwarning`
* `wc-cerror`