Add i18n resource cache buster
i18next caching plugin is able to use timed expiry. I have decided to
go with a querystring cachebuster using the same string as is used in
wc/resource/loader. This is because we cannot be sure that the use of
cache expiry is set correctly.

Happy to revisit this but I consider this fixes #1016.