Alleviate dialog jiggle
Dialogs may be set to autosize. When they do they resposiition
themselves after Ajax. If there are many concurrent ajax targets in a
dialog this may cause the dialog to reposition many times. To alleviate
this (but not to remove it completely as this is symptomatic of an
ill-considered dialog) I have created a test so that after ajax a
dialog will only reposition if it has no other aria-busy regions.