Move attribute helper XSLT templates
The XSLT templates in `wc.common.*` which output **only** an HTML
attribute have been moved to the file `wc.common.attribtues.xsl` which
was formerly `wc.common.attributeSets.xsl`. This is because these
cannot be overridden: they **must** output the attributes (or not)
based on the XML. This is part of #689 but also makes #639 easier.