Make XSLT non-optional
XSLT must take place in WComponents not in browser.

1. Deprecate config property and its accessor and set getter to always
return true. This is to prevent a non-backwards compatible API change
as the Config properties are all public.
2. Update the interceptor to remove calls to the config property and
**always** do XSLT (unless “magic” header set).
3. Update tests.
4. Update wcomponents-lde POM as there is a test dependency on the
theme. This unmet dependency used to cause an error which did not cause
the tests to fail as it resulted in a HTTP 404. With the theme XSLT
being required to be in the classPath to render, however, this missing
dependency was exposed.