WLabel refactor
Updated WLabelRenderer and schema to allow the renderer to give a hint
to ui:label about the labelled component. This has allowed the removal
of most of the remaining label to element lookups and has significantly
simplified the XSLT.

Split the schemas and XSLT for WFileWidget and WMultiFileWidget. This
helps with the WLabel XSLT simplification.

Override of WLabel method add(WComponent) so that if the component
being added is an Input the component will be set as the WLabel’s
labelled component.

Put wrapper around simple inputs and removed last WLabel element lookup.

Added two new marker interfaces “Labelable” and “InputGroup” used to
allow WLabelRenderer to more easily identify types of labelable
components. An Input may be both in which case it will be treated as an
input group by WLabelRenderer.