Simplify XSLT build
Most of the old XSLT build was there to support an unnecessarily
complex override mechanism and debug-only XSLT which we did not want to
send to the client (unless in debug mode). The override mechanism is no
longer relevant and the XSLT is no longer sent to the client so it does
not need to be so aggressively shrunk. This allows us to have a single
XSLT file, thereby simplifying the build.

This has flow-on simplifications to build-tests as XSLT is no longer
needed (or supported) in intern.