1158 double handlebars (#1169)
Partial fix for double handling handlebars: this is a 1.3 version of the solution for #1158.

Having tried XML output from the XSLT processor along with ensuring no
empty elements got short-tagged and `script` elements did not get
output escaping I remembered I converted the whole lot to XSLT2.0 so I
was able to use output method `xhtml`.

This has its own problems: script elements are no longer automatically
output escaped so JavaScript which contains `&&`, `<` etc can go
terribly wrong. To alleviate this I have built a template match on
`html:script` which will ensure the content is unescaped. Unfortunately
to do this just for script elements requires using the deprecated
`disable-output-escaping="on"` attribute on `xsl:value-of` and/or
`xsl:text` when doing the output. We use what we can.

This is an exceptionally good argument against XML/XSLT for web
applications.