Make client side tests run when `mvn test` invoked
This has been a weakness since we left JSTestDriver - client side unit
tests only run if saucelabs env variables are set otherwise they must
be manually invoked. This change causes the tests to run in headless
firefox as the default. The other two options are still available.
This makes firefox a fixture but I think that's acceptable for anyone
running client side unit tests. We could make this smarter later.