Finally make use of Observer.subscribe return val
As many comments in the source code attest, Observer.subscribe did not
return anything sensible or useful in any way. It was stupid to return
anything at all, API design lesson here is never return anything from
an API method unless you need it and it makes sense. It does not break
an API to add a return value later, but it does to change it.