Update event manager to allow recursive firing
The event manager used to only allow one event of any kind to fire
at a time, e.g. "click" would not fire while "change" was processing.
Then we relaxed it so that click would only prevent other clicks from
firing. Subsequent calls would be queued until the next event loop.
This was to prevent infinite event recursion but it comes at a cost.
I think we could completely do away with this protection but there is
a chance that bad code exists since we have been guarding it forever.
This change allows some recursion and I'm pretty sure it will allow
100% of legitimate cases to process synchronously whilst only pushing
potentially buggy code to the next event loop.