Upgrade ESLint to latest version
This was actually HUGE and that's why it wasn't done earlier.
It had to happen because:

- We were over three major versions behind 3.19.0 > 6.4.0
- Github was telling us "We found a potential security vulnerability
	in one of your dependencies." which was ESLint

Unfortunately ESLint seems to have changed its mind on a lot of things
in the meantime resulting in a LOT of refactoring with 18386 issues.

I let ESLint auto fix the test code but attended to the main
codebase myself.

I considered using Prettier instead but while it may produce consistent
results formatting involves a degree of artistic comprehension.
I haven't ruled out using Prettier in future. Will consider further.