Fix ServletUtil.extractCookie NPE with no cookies
HttpServletRequest.getCookies returns null when no cookies are set but
ServletUtil.extractCookie did not perform a null check and therefore
threw a NPE. While this was tested the tests assumed that the mock http
request MockHttpServletRequest was spec compliant. It was not since it
returned an empty collection instead of null. Fixes #1682