Add @FunctionalIterface to appropriate interfaces and classes
Adding @FunctionalInterface acts as documentation and assists the
compiler.

@FunctionalInterface compilation errors to be thrown just like
@Override. When @Override is added to a member and it doesn't override
anything then the compiler throws a compilation error. In the same way,
@FunctionalInterface will cause compilation errors if the class or
interface is not a functional interface.
It also acts as documentation for the developer, indication the an
instance can be created via a lambda expression, a method reference or a
constructor reference.