Restore public API of getTag to WDefinitionList
By not setting the tag property on items added for a term, then the
public API of WDefinitionList is changed. This should be avoided until
the next major version. So, in order to maintain this API the tag for is
set via setTag(), but the tag is not used internally.