Add mailcheck to email fields
Q. Why not use @zootools/email-spell-checker?
A. Because we want a client side checker. The new email-spell-checker
is currently designed to run server side, on a nodejs runtime.
The older mailcheck is highly configurable.

Note, this was painful to implement using AMD because mailcheck has
annoyingly hard-coded a define with the module name 'mailcheck'.
Perhaps we should fork it and fix it.