Port more modules to ES6
Also making a distinction between public API and initialise API which
is not really part of the API at all. Also get rid of the now pointles
one-shot contructors which do nothing more than create an useless level
of indentation. A LONG time ago they used to provide awesome private
scope for JS "modules" but we get that for free now.