Fix file input accepted bug from previous commit
Remodel the accept function to optionally take the input files in its
own array as a second parameter, i.e. not bound to the given file input
element. This is to enable validation of files if they are given
separately from the element, which is sometimes the case with
validate.mjs use.

The separate file array parameter is checked first. If not given, then
the files already inside the element are checked.

Includes basic unit tests for the new functionality. Also now passes
the tests that were failing from the previous commit because of the bug.